TST: Refactor reconstruction tests into smaller modules
