+++ tests/ptycho/test_probe.py
@@ -2,9 +2,7 @@ import os
 import unittest
 
 import cupy as cp
-import matplotlib
-matplotlib.use('Agg')
-from matplotlib import pyplot as plt
+import matplotlib.pyplot as plt
 import numpy as np
 import scipy.io
 import tike.ptycho.probe
