+++ tests/ptycho/test_ptycho.py
@@ -46,47 +46,31 @@
 # POSSIBILITY OF SUCH DAMAGE.                                             #
 # #########################################################################
 
-import matplotlib
-
-matplotlib.use('Agg')
-
-import bz2
 import lzma
 import os
 import pickle
 import unittest
-import warnings
 
 import numpy as np
-import cupy as cp
 
-import tike.ptycho
-from tike.ptycho.probe import ProbeOptions
-from tike.ptycho.position import PositionOptions
 from tike.ptycho.object import ObjectOptions
-from tike.communicators import Comm, MPIComm
-from tike.ptycho.solvers.options import (
-    _resize_fft,
-    _resize_spline,
-    _resize_cubic,
-    _resize_lanczos,
-    _resize_linear,
-)
+from tike.ptycho.probe import ProbeOptions
+import tike.ptycho
 import tike.random
+
+from .io import (
+    result_dir,
+    data_dir,
+    _save_ptycho_result,
+)
+from .templates import (
+    SiemensStarSetup,
+    ReconstructTwice,
+)
 
 __author__ = "Daniel Ching, Xiaodong Yu"
 __copyright__ = "Copyright (c) 2018, UChicago Argonne, LLC."
 __docformat__ = 'restructuredtext en'
-
-testdir = os.path.dirname(os.path.dirname(__file__))
-
-try:
-    from mpi4py import MPI
-    _mpi_size = MPI.COMM_WORLD.Get_size()
-    _mpi_rank = MPI.COMM_WORLD.Get_rank()
-except ModuleNotFoundError:
-    _mpi_size = 1
-    _mpi_rank = 0
 
 
 class TestPtychoUtils(unittest.TestCase):
@@ -94,7 +78,7 @@ class TestPtychoUtils(unittest.TestCase):
 
     def test_gaussian(self):
         """Check ptycho.gaussian for correctness."""
-        fname = os.path.join(testdir, 'data/ptycho_gaussian.pickle.lzma')
+        fname = os.path.join(data_dir, 'ptycho_gaussian.pickle.lzma')
         weights = tike.ptycho.probe.gaussian(15, rin=0.8, rout=1.0)
         if os.path.isfile(fname):
             with lzma.open(fname, 'rb') as file:
@@ -122,6 +106,7 @@ class TestPtychoUtils(unittest.TestCase):
 
 
 class TestPtychoSimulate(unittest.TestCase):
+    """Test the forward model for consistency."""
 
     def create_dataset(
         self,
@@ -191,7 +176,7 @@ class TestPtychoSimulate(unittest.TestCase):
 
     def setUp(self):
         """Load a dataset for reconstruction."""
-        dataset_file = os.path.join(testdir, 'data/ptycho_setup.pickle.lzma')
+        dataset_file = os.path.join(data_dir, 'ptycho_setup.pickle.lzma')
         if not os.path.isfile(dataset_file):
             self.create_dataset(dataset_file)
         with lzma.open(dataset_file, 'rb') as file:
@@ -217,74 +202,14 @@ class TestPtychoSimulate(unittest.TestCase):
         np.testing.assert_allclose(np.sqrt(data), np.sqrt(self.data), atol=1e-6)
 
 
-class TemplatePtychoRecon():
-
-    def setUp(self, filename='data/siemens-star-small.npz.bz2'):
-        """Load a dataset for reconstruction."""
-        dataset_file = os.path.join(testdir, filename)
-        with bz2.open(dataset_file, 'rb') as f:
-            archive = np.load(f)
-            self.scan = archive['scan'][0]
-            self.data = archive['data'][0]
-            self.probe = archive['probe'][0]
-        self.scan -= np.amin(self.scan, axis=-2) - 20
-        self.probe = tike.ptycho.probe.add_modes_cartesian_hermite(
-            self.probe, 5)
-        self.probe = tike.ptycho.probe.adjust_probe_power(self.probe)
-        self.probe = tike.ptycho.probe.orthogonalize_eig(self.probe)
-
-        with Comm(1, mpi=MPIComm) as comm:
-            mask = tike.cluster.by_scan_stripes(
-                self.scan,
-                n=comm.mpi.size,
-                fly=1,
-                axis=0,
-            )[comm.mpi.rank]
-            self.scan = self.scan[mask]
-            self.data = self.data[mask]
-
-    def init_params(self):
-        return tike.ptycho.PtychoParameters(
-            psi=np.full(
-                (600, 600),
-                dtype=np.complex64,
-                fill_value=np.complex64(0.5 + 0j),
-            ),
-            probe=self.probe,
-            scan=self.scan,
-        )
-
-    def template_consistent_algorithm(self, *, params={}):
-        """Check ptycho.solver.algorithm for consistency."""
-        device_per_rank = cp.cuda.runtime.getDeviceCount() // _mpi_size
-        base_device = device_per_rank * _mpi_rank
-        with cp.cuda.Device(base_device):
-            # Call twice to check that reconstruction continuation is correct
-            for _ in range(2):
-                params['parameters'] = tike.ptycho.reconstruct(
-                    **params,
-                    data=self.data,
-                    num_gpu=tuple(
-                        i + base_device for i in range(device_per_rank)),
-                    use_mpi=_mpi_size > 1,
-                )
-
-        print()
-        print('\n'.join(f'{c[0]:1.3e}'
-                        for c in params['parameters'].algorithm_options.costs))
-        return params['parameters']
-
-
-class TestPtychoAbsorption(TemplatePtychoRecon, unittest.TestCase):
+class TestPtychoAbsorption(SiemensStarSetup, unittest.TestCase):
     """Test various ptychography reconstruction methods for consistency."""
 
     def test_absorption(self):
         """Check ptycho.object.get_absorption_image for consistency."""
         try:
-            import matplotlib
-            matplotlib.use('Agg')
             from matplotlib import pyplot as plt
-            fname = os.path.join(testdir, 'result', 'ptycho', 'absorption')
+            fname = os.path.join(result_dir, 'absorption')
             os.makedirs(fname, exist_ok=True)
             plt.imsave(
                 f'{fname}/{0}-ampli.png',
@@ -298,13 +223,21 @@ class TestPtychoAbsorption(TemplatePtychoRecon, unittest.TestCase):
             pass
 
 
-class TestPtychoRecon(TemplatePtychoRecon, unittest.TestCase):
+class PtychoRecon(
+        ReconstructTwice,
+        SiemensStarSetup,
+):
     """Test various ptychography reconstruction methods for consistency."""
 
     post_name = ""
 
     def test_init(self):
-        params = self.init_params()
+        """Just test PtychoParameter initialization."""
+        params = tike.ptycho.PtychoParameters(
+            psi=self.psi,
+            probe=self.probe,
+            scan=self.scan,
+        )
         params.algorithm_options = tike.ptycho.AdamOptions(
             num_batch=5,
             num_iter=16,
@@ -312,70 +245,107 @@ class TestPtychoRecon(TemplatePtychoRecon, unittest.TestCase):
         params.probe_options = ProbeOptions()
         params.object_options = ObjectOptions()
         _save_ptycho_result(
-            params, f"{'mpi-' if _mpi_size > 1 else ''}init{self.post_name}")
+            params,
+            f"{'mpi-' if self.mpi_size > 1 else ''}init{self.post_name}",
+        )
 
     def test_consistent_adam_grad(self):
         """Check ptycho.solver.adam_grad for consistency."""
-        params = self.init_params()
-        params.algorithm_options = tike.ptycho.AdamOptions(
-            num_batch=5,
-            num_iter=16,
+        params = tike.ptycho.PtychoParameters(
+            psi=self.psi,
+            probe=self.probe,
+            scan=self.scan,
+            algorithm_options=tike.ptycho.AdamOptions(
+                num_batch=5,
+                num_iter=16,
+            ),
+            probe_options=ProbeOptions(),
+            object_options=ObjectOptions(),
         )
-        params.probe_options = ProbeOptions()
-        params.object_options = ObjectOptions()
         _save_ptycho_result(
-            self.template_consistent_algorithm(params={
-                'parameters': params,
-            },), f"{'mpi-' if _mpi_size > 1 else ''}adam_grad{self.post_name}")
+            self.template_consistent_algorithm(
+                data=self.data,
+                params=params,
+            ),
+            f"{'mpi-' if self.mpi_size > 1 else ''}adam_grad{self.post_name}",
+        )
 
     def test_consistent_cgrad(self):
         """Check ptycho.solver.cgrad for consistency."""
-        params = self.init_params()
-        params.algorithm_options = tike.ptycho.CgradOptions(
-            num_batch=5,
-            num_iter=16,
+        params = tike.ptycho.PtychoParameters(
+            psi=self.psi,
+            probe=self.probe,
+            scan=self.scan,
+            algorithm_options=tike.ptycho.CgradOptions(
+                num_batch=5,
+                num_iter=16,
+            ),
+            probe_options=ProbeOptions(),
+            object_options=ObjectOptions(),
         )
-        params.probe_options = ProbeOptions()
-        params.object_options = ObjectOptions()
         _save_ptycho_result(
-            self.template_consistent_algorithm(params={
-                'parameters': params,
-            },), f"{'mpi-' if _mpi_size > 1 else ''}cgrad{self.post_name}")
+            self.template_consistent_algorithm(
+                data=self.data,
+                params=params,
+            ),
+            f"{'mpi-' if self.mpi_size > 1 else ''}cgrad{self.post_name}",
+        )
 
     def test_consistent_lstsq_grad(self):
         """Check ptycho.solver.lstsq_grad for consistency."""
-        params = self.init_params()
-        params.algorithm_options = tike.ptycho.LstsqOptions(
-            num_batch=5,
-            num_iter=16,
+        params = tike.ptycho.PtychoParameters(
+            psi=self.psi,
+            probe=self.probe,
+            scan=self.scan,
+            algorithm_options=tike.ptycho.LstsqOptions(
+                num_batch=5,
+                num_iter=16,
+            ),
+            probe_options=ProbeOptions(use_adaptive_moment=True,),
+            object_options=ObjectOptions(use_adaptive_moment=True,),
         )
-        params.probe_options = ProbeOptions(use_adaptive_moment=True,)
-        params.object_options = ObjectOptions(use_adaptive_moment=True,)
         _save_ptycho_result(
-            self.template_consistent_algorithm(params={
-                'parameters': params,
-            },), f"{'mpi-' if _mpi_size > 1 else ''}lstsq_grad{self.post_name}")
+            self.template_consistent_algorithm(
+                data=self.data,
+                params=params,
+            ),
+            f"{'mpi-' if self.mpi_size > 1 else ''}lstsq_grad{self.post_name}")
 
     def test_consistent_lstsq_grad_compact(self):
         """Check ptycho.solver.lstsq_grad for consistency."""
-        params = self.init_params()
-        params.algorithm_options = tike.ptycho.LstsqOptions(
-            num_batch=5,
-            num_iter=16,
-            batch_method='compact',
+        params = tike.ptycho.PtychoParameters(
+            psi=self.psi,
+            probe=self.probe,
+            scan=self.scan,
+            algorithm_options=tike.ptycho.LstsqOptions(
+                num_batch=5,
+                num_iter=16,
+                batch_method='compact',
+            ),
+            probe_options=ProbeOptions(use_adaptive_moment=True,),
+            object_options=ObjectOptions(use_adaptive_moment=True,),
         )
-        params.probe_options = ProbeOptions(use_adaptive_moment=True,)
-        params.object_options = ObjectOptions(use_adaptive_moment=True,)
         _save_ptycho_result(
-            self.template_consistent_algorithm(params={
-                'parameters': params,
-            },),
-            f"{'mpi-' if _mpi_size > 1 else ''}lstsq_grad-compact{self.post_name}"
+            self.template_consistent_algorithm(
+                data=self.data,
+                params=params,
+            ),
+            f"{'mpi-' if self.mpi_size > 1 else ''}lstsq_grad-compact{self.post_name}"
         )
 
     def test_consistent_lstsq_grad_variable_probe(self):
         """Check ptycho.solver.lstsq_grad for consistency."""
-        params = self.init_params()
+        params = tike.ptycho.PtychoParameters(
+            psi=self.psi,
+            probe=self.probe,
+            scan=self.scan,
+            algorithm_options=tike.ptycho.LstsqOptions(
+                num_batch=5,
+                num_iter=16,
+            ),
+            probe_options=ProbeOptions(use_adaptive_moment=True,),
+            object_options=ObjectOptions(use_adaptive_moment=True,),
+        )
         probes_with_modes = min(3, params.probe.shape[-3])
         params.eigen_probe, params.eigen_weights = tike.ptycho.probe.init_varying_probe(
             params.scan,
@@ -383,18 +353,13 @@ class TestPtychoRecon(TemplatePtychoRecon, unittest.TestCase):
             num_eigen_probes=3,
             probes_with_modes=probes_with_modes,
         )
-        params.algorithm_options = tike.ptycho.LstsqOptions(
-            num_batch=5,
-            num_iter=16,
+        result = self.template_consistent_algorithm(
+            data=self.data,
+            params=params,
         )
-        params.probe_options = ProbeOptions(use_adaptive_moment=True,)
-        params.object_options = ObjectOptions(use_adaptive_moment=True,)
-        result = self.template_consistent_algorithm(params={
-            'parameters': params,
-        },)
         _save_ptycho_result(
             result,
-            f"{'mpi-' if _mpi_size > 1 else ''}lstsq_grad-variable-probe{self.post_name}",
+            f"{'mpi-' if self.mpi_size > 1 else ''}lstsq_grad-variable-probe{self.post_name}",
         )
         assert np.all(result.eigen_weights[..., 1:, probes_with_modes:] == 0), (
             "These weights should be unused/untouched "
@@ -402,451 +367,53 @@ class TestPtychoRecon(TemplatePtychoRecon, unittest.TestCase):
 
     def test_consistent_rpie(self):
         """Check ptycho.solver.rpie for consistency."""
-        params = self.init_params()
-        params.algorithm_options = tike.ptycho.RpieOptions(
-            num_batch=5,
-            num_iter=16,
+        params = tike.ptycho.PtychoParameters(
+            psi=self.psi,
+            probe=self.probe,
+            scan=self.scan,
+            algorithm_options=tike.ptycho.RpieOptions(
+                num_batch=5,
+                num_iter=16,
+            ),
+            probe_options=ProbeOptions(),
+            object_options=ObjectOptions(),
         )
-        params.probe_options = ProbeOptions()
-        params.object_options = ObjectOptions()
         _save_ptycho_result(
-            self.template_consistent_algorithm(params={
-                'parameters': params,
-            },), f"{'mpi-' if _mpi_size > 1 else ''}rpie{self.post_name}")
+            self.template_consistent_algorithm(
+                data=self.data,
+                params=params,
+            ),
+            f"{'mpi-' if self.mpi_size > 1 else ''}rpie{self.post_name}",
+        )
 
     def test_consistent_dm(self):
         """Check ptycho.solver.dm for consistency."""
-        params = self.init_params()
-        params.algorithm_options = tike.ptycho.DmOptions(
-            num_iter=16,
-            num_batch=5,
+        params = tike.ptycho.PtychoParameters(
+            psi=self.psi,
+            probe=self.probe,
+            scan=self.scan,
+            algorithm_options=tike.ptycho.DmOptions(
+                num_iter=16,
+                num_batch=5,
+            ),
+            probe_options=ProbeOptions(),
+            object_options=ObjectOptions(),
         )
-        params.probe_options = ProbeOptions()
-        params.object_options = ObjectOptions()
         _save_ptycho_result(
-            self.template_consistent_algorithm(params={
-                'parameters': params,
-            },), f"{'mpi-' if _mpi_size > 1 else ''}dm{self.post_name}")
-
-
-@unittest.skipIf(
-    _mpi_size > 1,
-    reason="MPI not implemented for online reconstruction.",
-)
-class TestPtychoOnline(TestPtychoRecon, unittest.TestCase):
-    """Test ptychography reconstruction when data is streaming."""
-
-    post_name = "-online"
-
-    def setUp(self, chunks=16) -> None:
-        TestPtychoRecon.setUp(self)
-        data = np.array_split(self.data, chunks, axis=0)
-        scan = np.array_split(self.scan, chunks, axis=0)
-        assert len(data) == chunks
-        assert len(scan) == chunks
-
-        self.data = data[0]
-        self.scan = scan[0]
-        self.data_more = data[1:]
-        self.scan_more = scan[1:]
-
-    def template_consistent_algorithm(self, *, params={}):
-
-        if _mpi_size > 1:
-            raise NotImplementedError()
-
-        with tike.ptycho.Reconstruction(**params, data=self.data) as context:
-            context.iterate(2)
-            for d, s in zip(self.data_more, self.scan_more):
-                context.append_new_data(
-                    new_data=d,
-                    new_scan=s,
-                )
-                context.iterate(2)
-        result = context.parameters
-        print()
-        print('\n'.join(f'{c[0]:1.3e}' for c in result.algorithm_options.costs))
-        return result
-
-
-class TestPtychoPosition(TemplatePtychoRecon, unittest.TestCase):
-    """Test various ptychography reconstruction methods position correction."""
-
-    post_name = "-position"
-
-    def setUp(self, filename='data/position-error-247.pickle.bz2'):
-        """Load a dataset for reconstruction.
-
-        This position correction test dataset was collected by Tao Zhou at the
-        Center for Nanoscale Materials Hard X-ray Nanoprobe
-        (https://www.anl.gov/cnm).
-        """
-        dataset_file = os.path.join(testdir, filename)
-        with bz2.open(dataset_file, 'rb') as f:
-            [
-                self.data,
-                self.scan,
-                self.scan_truth,
-                self.probe,
-            ] = pickle.load(f)
-
-        with Comm(1, mpi=MPIComm) as comm:
-            mask = tike.cluster.by_scan_stripes(
-                self.scan,
-                n=comm.mpi.size,
-                fly=1,
-                axis=0,
-            )[comm.mpi.rank]
-            self.scan = self.scan[mask]
-            self.scan_truth = self.scan_truth[mask]
-            self.data = self.data[mask]
-
-    def _save_position_error_variance(self, result, algorithm):
-        if result is None or _mpi_rank > 0:
-            return
-        try:
-            import matplotlib
-            matplotlib.use('Agg')
-            from matplotlib import pyplot as plt
-            import tike.view
-            fname = os.path.join(testdir, 'result', 'ptycho', f'{algorithm}')
-            os.makedirs(fname, exist_ok=True)
-
-            f = plt.figure(dpi=600)
-            plt.title(algorithm)
-            tike.view.plot_positions_convergence(
-                self.scan_truth,
-                result.position_options.initial_scan,
-                result.scan,
-            )
-            plt.savefig(os.path.join(fname, 'position-error.svg'))
-            plt.close(f)
-
-            f = plt.figure(dpi=600)
-            plt.title(algorithm)
-            plt.scatter(
-                np.linalg.norm(self.scan_truth - result.scan, axis=-1),
-                result.position_options.confidence[..., 0],
-            )
-            plt.xlabel('position error')
-            plt.ylabel('position confidence')
-            plt.savefig(os.path.join(fname, 'position-confidence.svg'))
-            plt.close(f)
-
-            f = plt.figure(dpi=600)
-            plt.title(algorithm)
-            plt.scatter(
-                result.position_options.initial_scan[..., 0],
-                result.position_options.initial_scan[..., 1],
-                marker='o',
-            )
-            plt.scatter(
-                result.scan[..., 0],
-                result.scan[..., 1],
-                marker='o',
-                color='red',
-                facecolor='None',
-            )
-            plt.scatter(
-                result.position_options.transform(
-                    result.position_options.initial_scan)[..., 0],
-                result.position_options.transform(
-                    result.position_options.initial_scan)[..., 1],
-                marker='x',
-            )
-            plt.legend(['initial', 'result', 'global'])
-            plt.savefig(os.path.join(fname, 'position-models.svg'))
-            plt.close(f)
-
-        except ImportError:
-            pass
-
-    def test_consistent_rpie_off(self):
-        """Check ptycho.solver.rpie position correction."""
-        algorithm = f"{'mpi-' if _mpi_size > 1 else ''}rpie-off{self.post_name}"
-        params = self.init_params()
-        params.algorithm_options = tike.ptycho.RpieOptions(
-            num_batch=5,
-            num_iter=16,
-        )
-        params.probe_options = ProbeOptions()
-        params.object_options = ObjectOptions()
-        result = self.template_consistent_algorithm(params={
-            'parameters': params,
-        },)
-        _save_ptycho_result(result, algorithm)
-
-    def test_consistent_rpie(self):
-        """Check ptycho.solver.rpie position correction."""
-        algorithm = f"{'mpi-' if _mpi_size > 1 else ''}rpie{self.post_name}"
-        params = self.init_params()
-        params.algorithm_options = tike.ptycho.RpieOptions(
-            num_batch=5,
-            num_iter=16,
-        )
-        params.position_options = PositionOptions(
-            self.scan,
-            use_adaptive_moment=True,
-            use_position_regularization=True,
-        )
-        params.probe_options = ProbeOptions()
-        params.object_options = ObjectOptions()
-        result = self.template_consistent_algorithm(params={
-            'parameters': params,
-        },)
-        _save_ptycho_result(result, algorithm)
-        self._save_position_error_variance(result, algorithm)
-
-    def test_consistent_lstsq_grad(self):
-        """Check ptycho.solver.lstsq_grad for consistency."""
-        algorithm = f"{'mpi-' if _mpi_size > 1 else ''}lstsq_grad{self.post_name}"
-        params = self.init_params()
-        params.algorithm_options = tike.ptycho.LstsqOptions(
-            num_batch=5,
-            num_iter=16,
-        )
-        params.position_options = PositionOptions(
-            self.scan,
-            use_adaptive_moment=True,
-            use_position_regularization=True,
-        )
-        params.probe_options = ProbeOptions()
-        params.object_options = ObjectOptions()
-        result = self.template_consistent_algorithm(params={
-            'parameters': params,
-        },)
-        _save_ptycho_result(result, algorithm)
-        self._save_position_error_variance(result, algorithm)
-
-
-@unittest.skipIf('TIKE_TEST_CI' in os.environ,
-                 reason="Just for user reference; not needed on CI.")
-class TestPtychoPositionReference(TestPtychoPosition, unittest.TestCase):
-    """Test various ptychography reconstruction methods position correction."""
-
-    post_name = '-position-ref'
-
-    def setUp(self, filename='data/position-error-247.pickle.bz2'):
-        """Load a dataset for reconstruction.
-
-        This position correction test dataset was collected by Tao Zhou at the
-        Center for Nanoscale Materials Hard X-ray Nanoprobe
-        (https://www.anl.gov/cnm).
-        """
-        dataset_file = os.path.join(testdir, filename)
-        with bz2.open(dataset_file, 'rb') as f:
-            [
-                self.data,
-                _,
-                self.scan,
-                self.probe,
-            ] = pickle.load(f)
-        self.scan_truth = self.scan
-
-        with Comm(1, mpi=MPIComm) as comm:
-            mask = tike.cluster.by_scan_stripes(
-                self.scan,
-                n=comm.mpi.size,
-                fly=1,
-                axis=0,
-            )[comm.mpi.rank]
-            self.scan = self.scan[mask]
-            self.data = self.data[mask]
-            self.scan_truth = self.scan_truth[mask]
-
-
-def _save_eigen_probe(output_folder, eigen_probe):
-    import matplotlib
-    matplotlib.use('Agg')
-    from matplotlib import pyplot as plt
-    flattened = []
-    for i in range(eigen_probe.shape[-4]):
-        probe = eigen_probe[..., i, :, :, :]
-        flattened.append(
-            np.concatenate(
-                probe.reshape((-1, *probe.shape[-2:])),
-                axis=1,
-            ))
-    flattened = np.concatenate(flattened, axis=0)
-    with warnings.catch_warnings():
-        warnings.filterwarnings("ignore", category=UserWarning)
-        plt.imsave(
-            f'{output_folder}/eigen-phase.png',
-            np.angle(flattened),
-            # The output of np.angle is locked to (-pi, pi]
-            cmap=plt.cm.twilight,
-            vmin=-np.pi,
-            vmax=np.pi,
-        )
-        plt.imsave(
-            f'{output_folder}/eigen-ampli.png',
-            np.abs(flattened),
-        )
-
-
-def _save_probe(output_folder, probe, algorithm):
-    import matplotlib
-    matplotlib.use('Agg')
-    from matplotlib import pyplot as plt
-    flattened = np.concatenate(
-        probe.reshape((-1, *probe.shape[-2:])),
-        axis=1,
-    )
-    with warnings.catch_warnings():
-        warnings.filterwarnings("ignore", category=UserWarning)
-        plt.imsave(
-            f'{output_folder}/probe-phase.png',
-            np.angle(flattened),
-            # The output of np.angle is locked to (-pi, pi]
-            cmap=plt.cm.twilight,
-            vmin=-np.pi,
-            vmax=np.pi,
-        )
-        plt.imsave(
-            f'{output_folder}/probe-ampli.png',
-            np.abs(flattened),
-        )
-    f = plt.figure()
-    tike.view.plot_probe_power(probe)
-    plt.semilogy()
-    plt.title(algorithm)
-    plt.savefig(f'{output_folder}/probe-power.svg')
-    plt.close(f)
-
-
-def _save_ptycho_result(result, algorithm):
-    if result is None or _mpi_rank > 0:
-        return
-    try:
-        import matplotlib
-        matplotlib.use('Agg')
-        from matplotlib import pyplot as plt
-        import tike.view
-        fname = os.path.join(testdir, 'result', 'ptycho', f'{algorithm}')
-        os.makedirs(fname, exist_ok=True)
-
-        fig = plt.figure()
-        ax1, ax2 = tike.view.plot_cost_convergence(
-            result.algorithm_options.costs,
-            result.algorithm_options.times,
-        )
-        ax2.set_xlim(0, 20)
-        ax1.set_ylim(10**(-1), 10**2)
-        fig.suptitle(algorithm)
-        fig.tight_layout()
-        plt.savefig(os.path.join(fname, 'convergence.svg'))
-        plt.close(fig)
-        plt.imsave(
-            f'{fname}/{0}-phase.png',
-            np.angle(result.psi).astype('float32'),
-            # The output of np.angle is locked to (-pi, pi]
-            cmap=plt.cm.twilight,
-            vmin=-np.pi,
-            vmax=np.pi,
-        )
-        plt.imsave(
-            f'{fname}/{0}-ampli.png',
-            np.abs(result.psi).astype('float32'),
-        )
-        import tifffile
-        tifffile.imwrite(
-            f'{fname}/{0}-ampli.tiff',
-            np.abs(result.psi).astype('float32'),
-        )
-        _save_probe(fname, result.probe, algorithm)
-        if result.eigen_weights is not None:
-            _save_eigen_weights(fname, result.eigen_weights)
-            if result.eigen_weights.shape[-2] > 1:
-                _save_eigen_probe(fname, result.eigen_probe)
-    except ImportError:
-        pass
-
-
-def _save_eigen_weights(fname, weights):
-    import matplotlib
-    matplotlib.use('Agg')
-    from matplotlib import pyplot as plt
-    plt.figure()
-    tike.view.plot_eigen_weights(weights)
-    plt.suptitle('weights')
-    plt.tight_layout()
-    plt.savefig(f'{fname}/weights.svg')
-
-
-@unittest.skipIf(_mpi_size > 1, reason="MPI not implemented for multi-grid.")
-class PtychoReconMultiGrid():
-    """Test ptychography multi-grid reconstruction method."""
-
-    def interp(self, x, f):
-        pass
-
-    def template_consistent_algorithm(self, *, params={}):
-        """Check ptycho.solver.algorithm for consistency."""
-
-        if _mpi_size > 1:
-            raise NotImplementedError()
-
-        device_per_rank = cp.cuda.runtime.getDeviceCount() // _mpi_size
-        base_device = device_per_rank * _mpi_rank
-        with cp.cuda.Device(base_device):
-            parameters = tike.ptycho.reconstruct_multigrid(
-                **params,
+            self.template_consistent_algorithm(
                 data=self.data,
-                num_gpu=tuple(i + base_device for i in range(device_per_rank)),
-                use_mpi=_mpi_size > 1,
-                num_levels=2,
-                interp=self.interp,
-            )
-
-        print()
-        print('\n'.join(
-            f'{c[0]:1.3e}' for c in parameters.algorithm_options.costs))
-        return parameters
+                params=params,
+            ),
+            f"{'mpi-' if self.mpi_size > 1 else ''}dm{self.post_name}",
+        )
 
 
-class TestPtychoReconMultiGridFFT(PtychoReconMultiGrid, TestPtychoRecon,
-                                  unittest.TestCase):
-
-    post_name = '-multigrid-fft'
-
-    def interp(self, x, f):
-        return _resize_fft(x, f)
-
-
-if False:
-    # Don't need to run these tests on CI every time.
-
-    class TestPtychoReconMultiGridLinear(PtychoReconMultiGrid, TestPtychoRecon,
-                                         unittest.TestCase):
-
-        post_name = '-multigrid-linear'
-
-        def interp(self, x, f):
-            return _resize_linear(x, f)
-
-    class TestPtychoReconMultiGridCubic(PtychoReconMultiGrid, TestPtychoRecon,
-                                        unittest.TestCase):
-
-        post_name = '-multigrid-cubic'
-
-        def interp(self, x, f):
-            return _resize_cubic(x, f)
-
-    class TestPtychoReconMultiGridLanczos(PtychoReconMultiGrid, TestPtychoRecon,
-                                          unittest.TestCase):
-
-        post_name = '-multigrid-lanczos'
-
-        def interp(self, x, f):
-            return _resize_lanczos(x, f)
-
-    class TestPtychoReconMultiGridSpline(PtychoReconMultiGrid, TestPtychoRecon,
-                                         unittest.TestCase):
-
-        post_name = '-multigrid-spline'
-
-        def interp(self, x, f):
-            return _resize_spline(x, f)
+class TestPtychoRecon(
+        PtychoRecon,
+        unittest.TestCase,
+):
+    """Separate test from implementation so that PtychoRecon can be imported elsewhere."""
+    pass
 
 
 if __name__ == '__main__':
